Set value IDs on commands and subcommands. Remove ParseArguments.hpp and ParseArguments.cpp.
