+++ single_src/Argos/Argos.cpp
@@ -505,7 +505,9 @@ namespace argos
      * Finish the initialization of this command and any subcommands, and
      * make them ready for parsing arguments.
      */
-    void finish_initialization(CommandData& cmd, const ParserData& data);
+    void finish_initialization(CommandData& cmd,
+                               const ParserData& data,
+                               ValueId start_id = ValueId(0));
 
     bool has_flag(const CommandData& cmd,
                   std::string_view flag,
@@ -1577,24 +1579,6 @@ namespace argos
 }
 
 //****************************************************************************
-// Copyright © 2024 Jan Erik Breimo. All rights reserved.
-// Created by Jan Erik Breimo on 2024-09-03.
-//
-// This file is distributed under the BSD License.
-// License text is included with the source distribution.
-//****************************************************************************
-
-namespace argos
-{
-    ParsedArguments parse_arguments(std::vector<std::string_view> args,
-                                    const std::shared_ptr<ParserData>& data);
-
-    ArgumentIterator
-    make_argument_iterator(std::vector<std::string_view> args,
-                           const std::shared_ptr<ParserData>& data);
-}
-
-//****************************************************************************
 // Copyright © 2020 Jan Erik Breimo. All rights reserved.
 // Created by Jan Erik Breimo on 2020-01-26.
 //
@@ -1660,6 +1644,22 @@ namespace argos
         }
 
         constexpr char DEFAULT_NAME[] = "UNINITIALIZED";
+
+        ParsedArguments parse_arguments(std::vector<std::string_view> args,
+                                        const std::shared_ptr<ParserData>& data)
+        {
+            finish_initialization(*data);
+            return ParsedArguments(
+                ArgumentIteratorImpl::parse(std::move(args), data));
+        }
+
+        ArgumentIterator
+        make_argument_iterator(std::vector<std::string_view> args,
+                               const std::shared_ptr<ParserData>& data)
+        {
+            finish_initialization(*data);
+            return {std::move(args), data};
+        }
     }
 
     ArgumentParser::ArgumentParser()
@@ -2809,10 +2809,10 @@ namespace argos
           section(rhs.section)
     {
         arguments.reserve(rhs.arguments.size());
-        for (const auto& a: rhs.arguments)
+        for (const auto& a : rhs.arguments)
             arguments.push_back(std::make_unique<ArgumentData>(*a));
         options.reserve(rhs.options.size());
-        for (const auto& o: rhs.options)
+        for (const auto& o : rhs.options)
             options.push_back(std::make_unique<OptionData>(*o));
     }
 
@@ -2842,17 +2842,17 @@ namespace argos
 
         arguments.clear();
         arguments.reserve(rhs.arguments.size());
-        for (const auto& a: rhs.arguments)
+        for (const auto& a : rhs.arguments)
             arguments.push_back(std::make_unique<ArgumentData>(*a));
 
         options.clear();
         options.reserve(rhs.options.size());
-        for (const auto& o: rhs.options)
+        for (const auto& o : rhs.options)
             options.push_back(std::make_unique<OptionData>(*o));
 
         commands.clear();
         commands.reserve(rhs.commands.size());
-        for (const auto& c: rhs.commands)
+        for (const auto& c : rhs.commands)
             commands.push_back(std::make_unique<CommandData>(*c));
 
         return *this;
@@ -2890,7 +2890,7 @@ namespace argos
 
         bool has_help_option(const CommandData& cmd)
         {
-            for (const auto& o: cmd.options)
+            for (const auto& o : cmd.options)
             {
                 if (o->type == OptionType::HELP)
                     return true;
@@ -2942,15 +2942,91 @@ namespace argos
         {
             add_help_option(cmd, settings);
         }
+
+        ValueId set_value_ids(const CommandData& cmd,
+                              ValueId start_id = ValueId(0))
+        {
+            struct InternalIdMaker
+            {
+                std::map<std::string_view, ValueId> explicit_ids;
+                ValueId n;
+
+                explicit InternalIdMaker(ValueId start_id)
+                    : n(start_id)
+                {
+                }
+
+                std::optional<ValueId> find_value_id(std::string_view name)
+                {
+                    const auto it = explicit_ids.find(name);
+                    if (it == explicit_ids.end())
+                        return {};
+                    return it->second;
+                }
+
+                ValueId make_value_id(std::string_view name)
+                {
+                    if (const auto id = find_value_id(name))
+                        return *id;
+                    n = ValueId(n + 1);
+                    explicit_ids.emplace(name, n);
+                    return n;
+                }
+
+                ValueId make_value_id(const std::vector<std::string>& names)
+                {
+                    for (const auto& name : names)
+                    {
+                        if (const auto id = find_value_id(name))
+                            return *id;
+                    }
+                    n = ValueId(n + 1);
+                    for (const auto& name : names)
+                        explicit_ids.emplace(name, n);
+                    return n;
+                }
+            };
+
+            InternalIdMaker id_maker(start_id);
+            for (const auto& a : cmd.arguments)
+            {
+                if (!a->value.empty())
+                {
+                    a->value_id = id_maker.make_value_id(a->value);
+                    id_maker.explicit_ids.emplace(a->name, a->value_id);
+                }
+                else
+                {
+                    a->value_id = id_maker.make_value_id(a->name);
+                }
+            }
+            for (const auto& o : cmd.options)
+            {
+                if (o->operation == OptionOperation::NONE)
+                    continue;
+                if (!o->alias.empty())
+                {
+                    o->value_id = id_maker.make_value_id(o->alias);
+                    for (auto& f : o->flags)
+                        id_maker.explicit_ids.emplace(f, o->value_id);
+                }
+                else
+                {
+                    o->value_id = id_maker.make_value_id(o->flags);
+                }
+            }
+
+            return id_maker.n;
+        }
     }
 
     bool has_flag(const CommandData& cmd,
                   std::string_view flag,
                   const ParserSettings& settings)
     {
-        for (auto& o: cmd.options)
+        for (auto& o : cmd.options)
         {
-            for (auto& f: o->flags)
+            for (auto& f : o->flags)
             {
                 if (are_equal(f, flag, settings.case_insensitive))
                     return true;
@@ -2959,12 +3035,15 @@ namespace argos
         return false;
     }
 
-    void finish_initialization(CommandData& cmd, const ParserData& data)
+    void finish_initialization(CommandData& cmd,
+                               const ParserData& data,
+                               ValueId start_id)
     {
         update_require_command(cmd);
         add_help_option(cmd, data.parser_settings);
-        for (auto& c: cmd.commands)
-            finish_initialization(*c, data);
+        start_id = set_value_ids(cmd, start_id);
+        for (auto& c : cmd.commands)
+            finish_initialization(*c, data, start_id);
     }
 }
 
@@ -3856,192 +3935,6 @@ namespace argos
 }
 
 //****************************************************************************
-// Copyright © 2024 Jan Erik Breimo. All rights reserved.
-// Created by Jan Erik Breimo on 2024-09-03.
-//
-// This file is distributed under the BSD License.
-// License text is included with the source distribution.
-//****************************************************************************
-
-namespace argos
-{
-    ParsedArguments parse_arguments(std::vector<std::string_view> args,
-                                    const std::shared_ptr<ParserData>& data)
-    {
-        finish_initialization(*data);
-        return ParsedArguments(
-            ArgumentIteratorImpl::parse(std::move(args), data));
-    }
-
-    ArgumentIterator
-    make_argument_iterator(std::vector<std::string_view> args,
-                           const std::shared_ptr<ParserData>& data)
-    {
-        finish_initialization(*data);
-        return {std::move(args), data};
-    }
-}
-
-//****************************************************************************
-// Copyright © 2020 Jan Erik Breimo. All rights reserved.
-// Created by Jan Erik Breimo on 2020-02-13.
-//
-// This file is distributed under the BSD License.
-// License text is included with the source distribution.
-//****************************************************************************
-
-#include <cstdlib>
-
-namespace argos
-{
-    namespace
-    {
-        template <typename T>
-        T str_to_int(const char* str, char** endp, int base);
-
-        template <>
-        long str_to_int<long>(const char* str, char** endp, int base)
-        {
-            return strtol(str, endp, base);
-        }
-
-        template <>
-        long long str_to_int<long long>(const char* str, char** endp, int base)
-        {
-            return strtoll(str, endp, base);
-        }
-
-        template <>
-        unsigned long
-        str_to_int<unsigned long>(const char* str, char** endp, int base)
-        {
-            return strtoul(str, endp, base);
-        }
-
-        template <>
-        unsigned long long
-        str_to_int<unsigned long long>(const char* str, char** endp, int base)
-        {
-            return strtoull(str, endp, base);
-        }
-
-        template <typename T>
-        std::optional<T> parse_integer_impl(const std::string& str, int base)
-        {
-            if (str.empty())
-                return {};
-            char* endp = nullptr;
-            errno = 0;
-            auto value = str_to_int<T>(str.c_str(), &endp, base);
-            if (endp == str.c_str() + str.size() && errno == 0)
-                return value;
-            return {};
-        }
-    }
-
-    template <>
-    std::optional<int> parse_integer<int>(const std::string& str, int base)
-    {
-        const auto n = parse_integer_impl<long>(str, base);
-        if (!n)
-            return {};
-
-        if constexpr (sizeof(int) != sizeof(long))
-        {
-            if (*n < INT_MIN || INT_MAX < *n)
-                return {};
-        }
-        return static_cast<int>(*n);
-    }
-
-    template <>
-    std::optional<unsigned>
-    parse_integer<unsigned>(const std::string& str, int base)
-    {
-        auto n = parse_integer_impl<unsigned long>(str, base);
-        if (!n)
-            return {};
-
-        if constexpr (sizeof(unsigned) != sizeof(unsigned long))
-        {
-            if (UINT_MAX < *n)
-                return {};
-        }
-        return static_cast<unsigned>(*n);
-    }
-
-    template <>
-    std::optional<long> parse_integer<long>(const std::string& str, int base)
-    {
-        return parse_integer_impl<long>(str, base);
-    }
-
-    template <>
-    std::optional<long long>
-    parse_integer<long long>(const std::string& str, int base)
-    {
-        return parse_integer_impl<long long>(str, base);
-    }
-
-    template <>
-    std::optional<unsigned long>
-    parse_integer<unsigned long>(const std::string& str, int base)
-    {
-        return parse_integer_impl<unsigned long>(str, base);
-    }
-
-    template <>
-    std::optional<unsigned long long>
-    parse_integer<unsigned long long>(const std::string& str, int base)
-    {
-        return parse_integer_impl<unsigned long long>(str, base);
-    }
-
-    namespace
-    {
-        template <typename T>
-        T str_to_float(const char* str, char** endp);
-
-        template <>
-        float str_to_float<float>(const char* str, char** endp)
-        {
-            return strtof(str, endp);
-        }
-
-        template <>
-        double str_to_float<double>(const char* str, char** endp)
-        {
-            return strtod(str, endp);
-        }
-
-        template <typename T>
-        std::optional<T> parse_floating_point_impl(const std::string& str)
-        {
-            if (str.empty())
-                return {};
-            char* endp = nullptr;
-            errno = 0;
-            auto value = str_to_float<T>(str.c_str(), &endp);
-            if (endp == str.c_str() + str.size() && errno == 0)
-                return value;
-            return {};
-        }
-    }
-
-    template <>
-    std::optional<float> parse_floating_point<float>(const std::string& str)
-    {
-        return parse_floating_point_impl<float>(str);
-    }
-
-    template <>
-    std::optional<double> parse_floating_point<double>(const std::string& str)
-    {
-        return parse_floating_point_impl<double>(str);
-    }
-}
-
-//****************************************************************************
 // Copyright © 2020 Jan Erik Breimo. All rights reserved.
 // Created by Jan Erik Breimo on 2020-01-26.
 //
@@ -4606,74 +4499,6 @@ namespace argos
 {
     namespace
     {
-        void set_value_ids(const ParserData& data)
-        {
-            struct InternalIdMaker
-            {
-                std::map<std::string_view, ValueId> explicit_ids;
-                ValueId n = ValueId(0);
-
-                std::optional<ValueId> find_value_id(std::string_view name)
-                {
-                    const auto it = explicit_ids.find(name);
-                    if (it == explicit_ids.end())
-                        return {};
-                    return it->second;
-                }
-
-                ValueId make_value_id(std::string_view name)
-                {
-                    if (const auto id = find_value_id(name))
-                        return *id;
-                    n = ValueId(n + 1);
-                    explicit_ids.emplace(name, n);
-                    return n;
-                }
-
-                ValueId make_value_id(const std::vector<std::string>& names)
-                {
-                    for (const auto& name: names)
-                    {
-                        if (const auto id = find_value_id(name))
-                            return *id;
-                    }
-                    n = ValueId(n + 1);
-                    for (const auto& name: names)
-                        explicit_ids.emplace(name, n);
-                    return n;
-                }
-            };
-
-            InternalIdMaker id_maker;
-            for (const auto& a: data.command.arguments)
-            {
-                if (!a->value.empty())
-                {
-                    a->value_id = id_maker.make_value_id(a->value);
-                    id_maker.explicit_ids.emplace(a->name, a->value_id);
-                }
-                else
-                {
-                    a->value_id = id_maker.make_value_id(a->name);
-                }
-            }
-            for (const auto& o: data.command.options)
-            {
-                if (o->operation == OptionOperation::NONE)
-                    continue;
-                if (!o->alias.empty())
-                {
-                    o->value_id = id_maker.make_value_id(o->alias);
-                    for (auto& f: o->flags)
-                        id_maker.explicit_ids.emplace(f, o->value_id);
-                }
-                else
-                {
-                    o->value_id = id_maker.make_value_id(o->flags);
-                }
-            }
-        }
-
         void add_version_option(ParserData& data)
         {
             if (data.version.empty())
@@ -4723,7 +4548,165 @@ namespace argos
     {
         add_version_option(data);
         finish_initialization(data.command, data);
-        set_value_ids(data);
+    }
+}
+
+//****************************************************************************
+// Copyright © 2020 Jan Erik Breimo. All rights reserved.
+// Created by Jan Erik Breimo on 2020-02-13.
+//
+// This file is distributed under the BSD License.
+// License text is included with the source distribution.
+//****************************************************************************
+
+#include <cstdlib>
+
+namespace argos
+{
+    namespace
+    {
+        template <typename T>
+        T str_to_int(const char* str, char** endp, int base);
+
+        template <>
+        long str_to_int<long>(const char* str, char** endp, int base)
+        {
+            return strtol(str, endp, base);
+        }
+
+        template <>
+        long long str_to_int<long long>(const char* str, char** endp, int base)
+        {
+            return strtoll(str, endp, base);
+        }
+
+        template <>
+        unsigned long
+        str_to_int<unsigned long>(const char* str, char** endp, int base)
+        {
+            return strtoul(str, endp, base);
+        }
+
+        template <>
+        unsigned long long
+        str_to_int<unsigned long long>(const char* str, char** endp, int base)
+        {
+            return strtoull(str, endp, base);
+        }
+
+        template <typename T>
+        std::optional<T> parse_integer_impl(const std::string& str, int base)
+        {
+            if (str.empty())
+                return {};
+            char* endp = nullptr;
+            errno = 0;
+            auto value = str_to_int<T>(str.c_str(), &endp, base);
+            if (endp == str.c_str() + str.size() && errno == 0)
+                return value;
+            return {};
+        }
+    }
+
+    template <>
+    std::optional<int> parse_integer<int>(const std::string& str, int base)
+    {
+        const auto n = parse_integer_impl<long>(str, base);
+        if (!n)
+            return {};
+
+        if constexpr (sizeof(int) != sizeof(long))
+        {
+            if (*n < INT_MIN || INT_MAX < *n)
+                return {};
+        }
+        return static_cast<int>(*n);
+    }
+
+    template <>
+    std::optional<unsigned>
+    parse_integer<unsigned>(const std::string& str, int base)
+    {
+        auto n = parse_integer_impl<unsigned long>(str, base);
+        if (!n)
+            return {};
+
+        if constexpr (sizeof(unsigned) != sizeof(unsigned long))
+        {
+            if (UINT_MAX < *n)
+                return {};
+        }
+        return static_cast<unsigned>(*n);
+    }
+
+    template <>
+    std::optional<long> parse_integer<long>(const std::string& str, int base)
+    {
+        return parse_integer_impl<long>(str, base);
+    }
+
+    template <>
+    std::optional<long long>
+    parse_integer<long long>(const std::string& str, int base)
+    {
+        return parse_integer_impl<long long>(str, base);
+    }
+
+    template <>
+    std::optional<unsigned long>
+    parse_integer<unsigned long>(const std::string& str, int base)
+    {
+        return parse_integer_impl<unsigned long>(str, base);
+    }
+
+    template <>
+    std::optional<unsigned long long>
+    parse_integer<unsigned long long>(const std::string& str, int base)
+    {
+        return parse_integer_impl<unsigned long long>(str, base);
+    }
+
+    namespace
+    {
+        template <typename T>
+        T str_to_float(const char* str, char** endp);
+
+        template <>
+        float str_to_float<float>(const char* str, char** endp)
+        {
+            return strtof(str, endp);
+        }
+
+        template <>
+        double str_to_float<double>(const char* str, char** endp)
+        {
+            return strtod(str, endp);
+        }
+
+        template <typename T>
+        std::optional<T> parse_floating_point_impl(const std::string& str)
+        {
+            if (str.empty())
+                return {};
+            char* endp = nullptr;
+            errno = 0;
+            auto value = str_to_float<T>(str.c_str(), &endp);
+            if (endp == str.c_str() + str.size() && errno == 0)
+                return value;
+            return {};
+        }
+    }
+
+    template <>
+    std::optional<float> parse_floating_point<float>(const std::string& str)
+    {
+        return parse_floating_point_impl<float>(str);
+    }
+
+    template <>
+    std::optional<double> parse_floating_point<double>(const std::string& str)
+    {
+        return parse_floating_point_impl<double>(str);
     }
 }
